% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newlm.R
\name{newlm}
\alias{newlm}
\title{Fit a linear regression model using the least squares method}
\usage{
newlm(formula, data)
}
\arguments{
\item{formula}{To specify the response and independent variables.}

\item{data}{The dataframe which contains the variables specified in the formula.}
}
\value{
Coefficients, fitted values and residuals,
various statistics such as the t-statistic, F-statistic, R-squared,
standard errors, confidence intervals, and the covariance matrix.
}
\description{
Fits a multiple linear regression model to the specified data using the least squares method.
Calculates model coefficients, fitted values, residuals, and provides various statistical measures,
including the t-statistic, F-statistic, R-squared. Additionally, computes
standard errors, confidence intervals, and the covariance matrix of the estimated coefficients.
}
\examples{
data(mtcars)
result <- newlm(mpg ~ wt + hp, mtcars)

}
